institutional_ai_trade_engine/pyproject.toml
[project]
name = "institutional-ai-trade-engine"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "python-telegram-bot>=21.4",
  "requests",
  "SQLAlchemy>=2.0",
  "apscheduler>=3.10",
  "ta>=0.11.0",
  "python-dotenv",
  "pytz",
  "gspread",
  "oauth2client"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"